clear dial form errors on change and refactor url normalize
